+++ lib/services/HttpCalls.dart
@@ -2,9 +2,10 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 import 'package:http/http.dart' as http;
-import '../models/response_model.dart';
+import 'package:my_widgets/models/response_model.dart';
 import '../my_widgets.dart';
 
 
@@ -25,17 +26,19 @@ class HttpCalls{
     return Uri.parse(sServerURL+postFix);
   }
 
-  static ViewResponse getDataObject(Response result) {
+  static dynamic getDataObject(Response result, {bool defaultResponse = true}) {
     Map<String, dynamic> userMap = jsonDecode(result.body);
-    ViewResponse response = ViewResponse.fromJson(userMap);
-    return response;
+    if(defaultResponse){
+      return ViewResponse.fromJson(userMap);
+    }
+    return userMap;
   }
 
 
 
 
-  static Future<ViewResponse> callGetApi(String endPoint,{bool hasAuth = true,required String token}) async {
-    ViewResponse response;
+  static Future<dynamic> callGetApi(String endPoint,{bool hasAuth = true,required String token, bool defaultResponse = true}) async {
+    dynamic response;
 
     Uri url = HttpCalls.getRequestURL(endPoint);
     print(url);
@@ -50,17 +53,21 @@ class HttpCalls{
     var result;
     try {
       result = await http.get(url,headers: header,).timeout(Duration(seconds: pTimeout));
-      print(result.body);
-      response = HttpCalls.getDataObject(result);
+      debugPrint('${result.body}');
+      response = HttpCalls.getDataObject(result, defaultResponse: defaultResponse);
+      
     } on TimeoutException catch (e) {
-      print("$e 001");
-      response = ViewResponse(message: e.toString(), status: false);
+      debugPrint("$e 001");
+      pShowToast(message: e.toString());
+
     } on HandshakeException catch (e) {
-      print("$e 002");
-      response = ViewResponse(message: e.toString(), status: false);
+      debugPrint("$e 002");
+      pShowToast(message: e.toString());
+
     } catch (e) {
-      print("Exception $e 003");
-      response = ViewResponse(message: e.toString(), status: false);
+      debugPrint("Exception $e 003");
+      pShowToast(message: e.toString());
+
 
 
     }
@@ -68,8 +75,8 @@ class HttpCalls{
     return response;
   }
 
-  static Future<ViewResponse> callPostApi(String endPoint,  Map params,{bool hasAuth = true,bool hasEncoded = true,required String token}) async {
-    ViewResponse response;
+  static Future<dynamic> callPostApi(String endPoint,  Map params,{bool hasAuth = true,bool hasEncoded = true,required String token, bool defaultResponse = true}) async {
+    dynamic response;
 
     print(params);
     Uri url = HttpCalls.getRequestURL(endPoint);
@@ -89,42 +96,43 @@ class HttpCalls{
       if (kDebugMode) {
         print(result.body);
       }
-      response = HttpCalls.getDataObject(result);
-      if (kDebugMode) {
-        print('response: $response');
-        print(response.data);
-        print(response.message);
-      }
+      response = HttpCalls.getDataObject(result, defaultResponse: defaultResponse);
+
     } on TimeoutException catch (e) {
       if (kDebugMode) {
         print("$e 001");
+        pShowToast(message: e.toString());
       }
-      response = ViewResponse( message: e.toString(), status: false);
     } on HandshakeException catch (e) {
       if (kDebugMode) {
         print("$e 002");
+        pShowToast(message: e.toString());
       }
-      response = ViewResponse(message: e.toString(), status: false);
     } catch (e) {
       if (kDebugMode) {
         print("Exception $e 003");
+        pShowToast(message: e.toString());
       }
       String error = e.toString();
       if(e.toString().contains('SocketException')){
         error = 'seems like internet issue';
+        pShowToast(message: error);
       }
-      response = ViewResponse(message: error, status: false);
     }
     return response;
   }
 
-  static Future<ViewResponse> callPutApi(String endPoint,  Map params,{bool hasAuth = true,bool hasEncoded = true,required String token}) async {
-    ViewResponse response;
+  static Future<dynamic> callPutApi(String endPoint,  Map params,{bool hasAuth = true,bool hasEncoded = true,required String token, bool defaultResponse = true}) async {
+    dynamic response;
 
 
-    print(params);
+    if (kDebugMode) {
+      print(params);
+    }
     Uri url = HttpCalls.getRequestURL(endPoint);
-    print(url);
+    if (kDebugMode) {
+      print(url);
+    }
     var header = {
       HttpHeaders.contentTypeHeader: 'application/json'
     };
@@ -137,30 +145,32 @@ class HttpCalls{
     try {
 
       result = await http.put(url,headers: header, body: utf8.encode(json.encode(params))).timeout(Duration(seconds: pTimeout));
-      print(result.body);
-      response = HttpCalls.getDataObject(result);
-      print('response: $response');
-      print(response.data);
-      print(response.message);
+      debugPrint('${result.body}');
+      response = HttpCalls.getDataObject(result, defaultResponse: defaultResponse);
+      
     } on TimeoutException catch (e) {
-      print("$e 001");
-      response = ViewResponse( message: e.toString(), status: false);
+      debugPrint("$e 001");
+      pShowToast(message: e.toString());
     } on HandshakeException catch (e) {
-      print("$e 002");
-      response = ViewResponse(message: e.toString(), status: false);
+      debugPrint("$e 002");
+      pShowToast(message: e.toString());
     } catch (e) {
-      print("Exception $e 003");
-      response = ViewResponse(message: e.toString(), status: false);
+      debugPrint("Exception $e 003");
+      pShowToast(message: e.toString());
     }
     return response;
   }
 
-  static Future<ViewResponse> callDeleteApi(String endPoint,  Map params,{bool hasAuth = true,bool hasEncoded = true, required String token}) async {
-    ViewResponse response;
+  static Future<dynamic> callDeleteApi(String endPoint,  Map params,{bool hasAuth = true,bool hasEncoded = true, required String token, bool defaultResponse = true}) async {
+    dynamic response;
 
-    print(params);
+    if (kDebugMode) {
+      print(params);
+    }
     Uri url = HttpCalls.getRequestURL(endPoint);
-    print(url);
+    if (kDebugMode) {
+      print(url);
+    }
     var header = {
       HttpHeaders.contentTypeHeader: 'application/json'
     };
@@ -175,26 +185,21 @@ class HttpCalls{
       request.body = jsonEncode(params);
       var streamedResponse = await request.send();
       var result = await Response.fromStream(streamedResponse);
-      print(result.body);
-      response = HttpCalls.getDataObject(result);
-      print('response: $response');
-      print(response.data);
-      print(response.message);
+      debugPrint('${result.body}');
+      response = HttpCalls.getDataObject(result, defaultResponse: defaultResponse);
+      
 
     } on TimeoutException catch (e) {
-      print("$e 001");
-      response = ViewResponse( message: e.toString(), status: false);
+      pShowToast(message: e.toString());
     } on HandshakeException catch (e) {
-      print("$e 002");
-      response = ViewResponse(message: e.toString(), status: false);
+      pShowToast(message: e.toString());
     } catch (e) {
-      print("Exception $e 003");
-      response = ViewResponse(message: e.toString(), status: false);
+      pShowToast(message: e.toString());
     }
     return response;
   }
 
-  static Future<ViewResponse> uploadFile(String endPoint, String filename, {String fileKey = 'image',bool isUserAvatar = false,bool hasAuth = true,Map<String, String>? params,required String token}) async {
+  static Future<dynamic> uploadFile(String endPoint, String filename, {String fileKey = 'image',bool isUserAvatar = false,bool hasAuth = true,Map<String, String>? params,required String token, bool defaultResponse = true}) async {
     Uri url = HttpCalls.getRequestURL(endPoint);
     print(url);
     var header = {
@@ -208,21 +213,69 @@ class HttpCalls{
       var request = MultipartRequest('POST', url,);
       request.files.add(await MultipartFile.fromPath(fileKey, filename));
       request.headers.addAll(header);
-      if(params != null)
+      if(params != null) {
         request.fields.addAll(params);
+      }
 
 
       var streamedResponse = await request.send();
       var result = await Response.fromStream(streamedResponse);
       print(result.body);
-      ViewResponse response = HttpCalls.getDataObject(result);
+      dynamic response = HttpCalls.getDataObject(result, defaultResponse: defaultResponse);
       return response;
 
     }catch (e){
       print(e.toString());
       pShowToast(message: e.toString());
-      return ViewResponse(status: false, message: 'Something went wrong, please try again in few minuets');
+      return null;
     }
+  }
+
+  static Future<dynamic> uploadImage(String filename,String fileType, {bool isUserAvatar = false,bool hasAuth = true, String thumbnail = '',required String token,required String userName, bool defaultResponse = true}) async {
+    dynamic file;
+    Uri url = HttpCalls.getRequestURL('file-upload');
+    print(url.toString());
+    var header = {
+      HttpHeaders.contentTypeHeader: 'application/json'
+    };
+
+    if(hasAuth) {
+      header[HttpHeaders.authorizationHeader] = 'Bearer $token';
+    }
+
+    try{
+      var request = http.MultipartRequest('POST', url,);
+      request.files.add(await http.MultipartFile.fromPath('file', filename));
+      if(fileType == 'video') {
+        request.files.add(await http.MultipartFile.fromPath('thumbnail', thumbnail));
+      }
+      request.fields['fileType'] = fileType;
+      request.fields['userName'] = userName;
+      if (kDebugMode) {
+        print(request.fields.toString());
+        print(request.files.toString());
+        print(request.files.toString());
+      }
+
+      if(hasAuth)request.headers.addAll(header);
+      var streamedResponse = await request.send();
+      var result = await Response.fromStream(streamedResponse);
+      if (kDebugMode) {
+        print("result.body");
+      }
+      dynamic response = HttpCalls.getDataObject(result, defaultResponse: defaultResponse);
+
+      if (response.status) {
+        file = response.data;
+      } else {
+        if (kDebugMode) {
+          print(response.message);
+        }
+      }
+    }catch (e){
+      pShowToast(message: e.toString());
+    }
+    return file;
   }